authorization for protected routes

+++ src/app/api/budget/[id]/delete.js
@@ -1,10 +1,18 @@
 import Budget from "@/models/Budget";
+import { getTokenFromRequest } from "@/utils/authHelper";
 import { ConnectToDB } from "@/utils/connect";
 
 export const DELETE = async (req, { params }) => {
   try {
     const { id } = params;
     await ConnectToDB();
+    const token = await getTokenFromRequest(req);
+
+    if (!token) {
+      return new Response(JSON.stringify({ error: "Unauthorized" }), {
+        status: 401,
+      });
+    }
     const deleteBudget = await Budget.findByIdAndDelete(id);
 
     if (!deleteBudget) {
